+++ libft/ft_lstclear.c
@@ -1,23 +1,31 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   builtin_pwd.c                                      :+:      :+:    :+:   */
+/*   ft_lstclear.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: wchae <wchae@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/06/28 00:14:31 by wchae             #+#    #+#             */
-/*   Updated: 2022/07/09 16:49:31 by wchae            ###   ########.fr       */
+/*   Created: 2022/07/09 17:06:34 by wchae             #+#    #+#             */
+/*   Updated: 2022/07/09 17:06:41 by wchae            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#include "libft.h"
 
-int	ft_pwd(void)
+void	ft_lstclear(t_list **lst, void (*del)(void *))
 {
-	char	*buf;
+	t_list	*next;
 
-	buf = getcwd(NULL, 0);
-	printf("%s\n", buf);
-	free (buf);
-	return (0);
+	if (!lst || !*lst || !del)
+		return ;
+	while (*lst)
+	{
+		next = (*lst)->next;
+		del((*lst)->data);
+		(*lst)->data = NULL;
+		free(*lst);
+		*lst = NULL;
+		*lst = next;
+	}
+	*lst = NULL;
 }
