+++ srcs/handle_builtin.c
@@ -1,16 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   exe_handler.c                                      :+:      :+:    :+:   */
+/*   handle_builtin.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: seseo <seseo@student.42.fr>                +#+  +:+       +#+        */
+/*   By: wchae <wchae@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/28 00:05:09 by wchae             #+#    #+#             */
-/*   Updated: 2022/07/09 15:39:43 by seseo            ###   ########.fr       */
+/*   Updated: 2022/07/09 16:54:02 by wchae            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../minishell.h"
+#include "minishell.h"
 
 int	check_builtin_cmd(t_token *tokens)
 {
