+++ srcs/settings.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   settings.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: seseo <seseo@student.42.fr>                +#+  +:+       +#+        */
+/*   By: wchae <wchae@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/27 23:53:43 by wchae             #+#    #+#             */
-/*   Updated: 2022/07/08 18:39:29 by seseo            ###   ########.fr       */
+/*   Updated: 2022/07/09 16:44:05 by wchae            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../minishell.h"
+#include "minishell.h"
 
 static t_env	*env_set(char	**envp)
 {
@@ -22,21 +22,13 @@ static t_env	*env_set(char	**envp)
 	{
 		eq = ft_strchr(*envp, '=');
 		if (eq)
-			env_lstadd_back(&env_list, ft_substr(*envp, 0, eq - *envp), ft_substr(*envp, eq - *envp + 1, ft_strlen(eq + 1)));
+			env_lstadd_back(&env_list,
+				ft_substr(*envp, 0, eq - *envp),
+				ft_substr(*envp, eq - *envp + 1,
+					ft_strlen(eq + 1)));
 		envp++;
 	}
 	return (env_list);
-}
-
-t_env	*find_env_node(t_env *env_list, char *key)
-{
-	while (env_list)
-	{
-		if (ft_strncmp(env_list->key, key, -1) == 0)
-			return (env_list);
-		env_list = env_list->next;
-	}
-	return (NULL);
 }
 
 void	set_env_node(t_env **env, char *key, char *val)
@@ -55,7 +47,7 @@ void	set_env_node(t_env **env, char *key, char *val)
 		env_lstadd_back(env, key, val);
 }
 
-void	env_init(t_env **env)
+static void	env_init(t_env **env)
 {
 	t_env		*shlvl;
 	int			lvl;
@@ -74,7 +66,6 @@ void	env_init(t_env **env)
 
 void	init_set(t_set *set, t_env **env)
 {
-	//env "=" 기준으로 split
 	ft_memset(set, 0, sizeof(t_set));
 	g_status = 0;
 	set->org_stdin = dup(STDIN_FILENO);
@@ -86,39 +77,3 @@ void	init_set(t_set *set, t_env **env)
 	*env = env_set(environ);
 	env_init(env);
 }
-
-// static int	ft_env_lstsize(t_env	*lst)
-// {
-// 	int	i;
-
-// 	i = 0;
-// 	while (lst)
-// 	{
-// 		lst = lst->next;
-// 		i++;
-// 	}
-// 	return (i);
-// }
-
-// static char	**convert_env_lst_to_dp(t_env	*env)
-// {
-// 	int		i;
-// 	int		len;
-// 	char	**envp;
-// 	char	*join_env;
-// 	char	*tmp;
-
-// 	len = ft_env_lstsize(env);
-// 	envp = (char **)malloc(sizeof(char *) * len + 1);
-// 	envp[len] = NULL;
-// 	i = 0;
-// 	while (len--)
-// 	{
-// 		tmp = ft_strjoin(env->key, "=");
-// 		join_env = ft_strjoin(tmp, env->value);
-// 		free(tmp);
-// 		envp[i++] = join_env;
-// 		env = env->next;
-// 	}
-// 	return (envp);
-// }
