+++ srcs/test_print.c
@@ -6,7 +6,7 @@
 /*   By: wchae <wchae@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/09 16:29:16 by wchae             #+#    #+#             */
-/*   Updated: 2022/07/09 16:55:16 by wchae            ###   ########.fr       */
+/*   Updated: 2022/07/09 17:04:16 by wchae            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
