+++ srcs/utils_error.c
@@ -1,0 +1,52 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   utils_error.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: wchae <wchae@student.42.fr>                +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/07/09 17:12:26 by wchae             #+#    #+#             */
+/*   Updated: 2022/07/09 17:19:41 by wchae            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+int	error_msg(char *msg)
+{
+	g_status = 2;
+	write(2, "bash: ", 6);
+	if (!msg)
+		write(2, "syntax error near unexpected token 'newline'", 44);
+	else if (msg[0] == '|' || msg[0] == '<' || msg[0] == '>')
+	{
+		write(2, "syntax error near unexpected token '", 36);
+		write(2, msg, ft_strlen(msg));
+		write(2, "'\n", 2);
+		return (ERROR);
+	}
+	else
+	{
+		g_status = 127;
+		write(2, msg, ft_strlen(msg));
+		if (ft_strncmp(strerror(errno), "Bad address", 12) == 0)
+			write(2, ": command not found", 19);
+		else
+		{
+			write(2, ": ", 2);
+			write(2, strerror(errno), ft_strlen(strerror(errno)));
+		}
+	}
+	write(2, "\n", 1);
+	return (g_status);
+}
+
+int	fd_print_err(int fd)
+{
+	if (fd == -1)
+	{
+		error_msg(strerror(errno));
+		return (EXIT_FAILURE);
+	}
+	return (EXIT_SUCCESS);
+}
